requirements.txt
pytest
PyMySQL
requests
Flask
python-dotenv
Flask-Migrate
Flask-SQLAlchemy
